Migrate build system to Bazel (Bazel 9 + Bzlmod)
- Add WORKSPACE, MODULE.bazel (rules_cc 0.2.17, platforms 1.0.0)
- Add BUILD.bazel for d2xx_mpsse library and 5 test binaries
- Add third_party/ftd2xx/BUILD.bazel with platform-select cc_import
  - Windows: interface_library (ftd2xx.lib, system_provided DLL)
  - Linux: static_library (libftd2xx.a)
- Add .bazelrc: C++20, -Wall -Wextra -Werror per platform
- Add .bazelignore: exclude bazel-* symlinks for cross-platform builds
- Fix narrowing conversion in MpsseDevice::readPins() (GCC -Werror=narrowing)
- Update .gitignore: add bazel-* output directories
- Update README: Bazel build instructions (recommended)

Build verified: Windows (MSVC, bazelisk build //...) + WSL2 GCC 13 PASS
Tests verified: all tests PASS on real FT2232H hardware (both platforms)

+++ src/MpsseDevice.cpp
@@ -291,7 +291,7 @@ void MpsseDevice::flush() {
 
 // ---------------------------------------------------------------------------
 uint8_t MpsseDevice::readPins(bool highByte) {
-    const uint8_t cmd[2] = { highByte ? 0x83u : 0x81u, 0x87u };
+    const uint8_t cmd[2] = { static_cast<uint8_t>(highByte ? 0x83u : 0x81u), 0x87u };
     write(std::span<const uint8_t>(cmd));
     return read(1)[0];
 }
